feat: chage id type Long to BigInteger

+++ src/main/kotlin/smalltalk/backend/domain/room/Room.kt
@@ -2,10 +2,11 @@ package smalltalk.backend.domain.room
 
 import com.fasterxml.jackson.annotation.JsonCreator
 import com.fasterxml.jackson.annotation.JsonProperty
+import java.math.BigInteger
 
 class Room @JsonCreator constructor(
     @JsonProperty("id")
-    val id: Long?,
+    val id: BigInteger?,
     @JsonProperty("name")
     val name: String?,
     @JsonProperty("idQueue")
